// src/dorc-web/src/apis/dorc-api/models/UserApiModel.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface UserApiModel
 */
export interface UserApiModel {
  /**
   * @type {number}
   * @memberof UserApiModel
   */
  Id?: number;
  /**
   * @type {string}
   * @memberof UserApiModel
   */
  LoginId?: string | null;
  /**
   * @type {string}
   * @memberof UserApiModel
   */
  DisplayName?: string | null;
  /**
   * @type {string}
   * @memberof UserApiModel
   */
  LoginType?: string | null;
  /**
   * @type {string}
   * @memberof UserApiModel
   */
  LanId?: string | null;
  /**
   * @type {string}
   * @memberof UserApiModel
   */
  Team?: string | null;
  /**
   * @type {string}
   * @memberof UserApiModel
   */
  LanIdType?: string | null;
}