// src/dorc-web/src/apis/dorc-api/models/GroupSearchResult.ts
// tslint:disable
/**
 * Dorc.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * @export
 * @interface GroupSearchResult
 */
export interface GroupSearchResult {
    /**
     * @type {string}
     * @memberof GroupSearchResult
     */
    DisplayName?: string | null;
    /**
     * @type {string}
     * @memberof GroupSearchResult
     */
    FullLogonName?: string | null;
}